Move utility functoins to lib folder
